--- RenderImage/RenderImage/Portable/MainPage.xaml
@@ -1,87 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:input="clr-namespace:Telerik.XamarinForms.Input;assembly=Telerik.XamarinForms.Input"
-             xmlns:barcode="clr-namespace:Telerik.XamarinForms.Barcode;assembly=Telerik.XamarinForms.Barcode"
-             x:Class="RenderImage.Portable.MainPage"
-             Title="Render Demo">
-
-    <ContentPage.ToolbarItems>
-        <ToolbarItem Icon="FullCrop.png"
-                     Text="Render All"
-                     Clicked="RenderAll_OnClicked" />
-        <ToolbarItem Icon="AbsCrop.png"
-                     Text="Abs Crop"
-                     Clicked="RenderCropAbsolute_OnClicked" />
-        <ToolbarItem Icon="PropCrop.png" 
-                     Text="Prop Crop"
-                     Clicked="RenderCropProportional_OnClicked" />
-    </ContentPage.ToolbarItems>
-
-    <Grid BackgroundColor="WhiteSmoke">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-
-        <Grid BackgroundColor="LightGray">
-            <barcode:RadBarcode x:Name="barcode"
-                                Value="https://www.telerik.com/xamarin-ui">
-                <barcode:RadBarcode.Symbology>
-                    <barcode:QRCode SizingMode="Stretch" />
-                </barcode:RadBarcode.Symbology>
-            </barcode:RadBarcode>
-        </Grid>
-
-        <StackLayout Grid.Row="1"
-                     Padding="5"
-                     Spacing="5">
-            <Grid>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition />
-                    <ColumnDefinition />
-                    <ColumnDefinition />
-                    <ColumnDefinition />
-                </Grid.ColumnDefinitions>
-
-                <input:RadEntry x:Name="XEntry"
-                                WatermarkText="X"
-                                Grid.Column="0">
-                    <input:RadEntry.BorderStyle>
-                        <input:BorderStyle BorderThickness="1"
-                                           BorderColor="DarkGreen" />
-                    </input:RadEntry.BorderStyle>
-                </input:RadEntry>
-                <input:RadEntry x:Name="YEntry"
-                                WatermarkText="Y"
-                                Grid.Column="1">
-                    <input:RadEntry.BorderStyle>
-                        <input:BorderStyle BorderThickness="1"
-                                           BorderColor="DarkGreen" />
-                    </input:RadEntry.BorderStyle>
-                </input:RadEntry>
-                <input:RadEntry x:Name="WidthEntry"
-                                WatermarkText="Width"
-                                Grid.Column="2">
-                    <input:RadEntry.BorderStyle>
-                        <input:BorderStyle BorderThickness="1"
-                                           BorderColor="CornflowerBlue" />
-                    </input:RadEntry.BorderStyle>
-                </input:RadEntry>
-                <input:RadEntry x:Name="HeightEntry"
-                                WatermarkText="Height"
-                                Grid.Column="3">
-                    <input:RadEntry.BorderStyle>
-                        <input:BorderStyle BorderThickness="1"
-                                           BorderColor="CornflowerBlue" />
-                    </input:RadEntry.BorderStyle>
-                </input:RadEntry>
-            </Grid>
-
-            <input:RadButton Text="Auto-fill Barcode Dimensions"
-                             TextColor="DarkRed"
-                             BackgroundColor="White"
-                             Clicked="ReadBarcodeDimensions_OnClicked" />
-        </StackLayout>
-    </Grid>
-</ContentPage>
