--- RenderImage/RenderImage/Portable/Models/MyPdfContent.cs
@@ -1,15 +1,0 @@
-﻿namespace RenderImage.Portable.Models
-{
-    public class MyPdfContent
-    {
-        public string Title { get; set; }
-
-        public string Body { get; set; }
-
-        public string BackgroundColor { get; set; }
-
-        public string RequestedFileName { get; set; }
-
-        public string ImageBase64 { get; set; }
-    }
-}
