--- RenderImage/RenderImage/Portable/ResultsPage.xaml
@@ -1,52 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:primitives="clr-namespace:Telerik.XamarinForms.Primitives;assembly=Telerik.XamarinForms.Primitives"
-             xmlns:input="clr-namespace:Telerik.XamarinForms.Input;assembly=Telerik.XamarinForms.Input"
-             x:Class="RenderImage.Portable.ResultsPage">
-
-    <Grid BackgroundColor="DarkSlateGray">
-        <Image x:Name="image"
-               Margin="20" />
-
-        <StackLayout VerticalOptions="End" Spacing="5">
-            <Label Text="PDF Generation Options"
-                   HorizontalTextAlignment="Center"/>
-            <input:RadButton x:Name="LocalGenerationButton"
-                                    Text="Generate Locally"  
-                                    BorderThickness="2" 
-                                    BorderColor="#4488F6" 
-                                    Clicked="LocalGenerationButton_OnClicked"/>
-
-            <input:RadButton x:Name="RemoteGenerationButton"
-                                    Text="Generate Remotely"  
-                                    BorderThickness="2" 
-                                    BorderColor="#4488F6" 
-                                    Clicked="RemoteGenerationButton_OnClicked"/>
-        </StackLayout>
-
-        <primitives:RadBusyIndicator x:Name="BusyIndicator"
-                                     AnimationType="Animation3"
-                                     AnimationContentColor="Blue"
-                                     BackgroundColor="#DDFFFFFF" />
-
-        <primitives:RadPopup.Popup>
-            <primitives:RadPopup x:Name="MessagePopup"
-                                 IsModal="True" 
-                                 Placement="Center"
-                                 OutsideBackgroundColor="#6F000000">
-                <primitives:RadBorder CornerRadius="15" 
-                                      BackgroundColor="Wheat"
-                                      Padding="10">
-                    <StackLayout Spacing="5">
-                        <Label x:Name="ErrorLabel"
-                               LineBreakMode="WordWrap"/>
-                        <Button Text="close"
-                                HorizontalOptions="Center"
-                                Clicked="ClosePopupButton_OnClicked"/>
-                    </StackLayout>
-                </primitives:RadBorder>
-            </primitives:RadPopup>
-        </primitives:RadPopup.Popup>
-    </Grid>
-</ContentPage>
