+++ RenderImage/RenderImage/Portable/Theme.xaml.cs
@@ -1,0 +1,12 @@
+﻿using Xamarin.Forms;
+
+namespace RenderImage.Portable
+{
+    public partial class Theme : ResourceDictionary
+    {
+        public Theme()
+        {
+            InitializeComponent();
+        }
+    }
+}
