+++ RenderImage/RenderImage/Portable/Views/DocumentViewerPage.xaml.cs
@@ -1,0 +1,20 @@
+﻿using Telerik.XamarinForms.PdfViewer;
+using Xamarin.Forms;
+
+namespace RenderImage.Portable.Views
+{
+    public partial class DocumentViewerPage : ContentPage
+    {
+		public DocumentViewerPage()
+		{
+			InitializeComponent();
+		}
+
+        protected override void OnAppearing()
+        {
+            Viewer.Source = new FileDocumentSource(App.PdfFilePath);
+
+            base.OnAppearing();
+        }
+    }
+}
