+++ RenderImage/RenderImage/UWP/RenderImage.UWP.csproj
@@ -102,21 +102,30 @@
     <AppxManifest Include="Package.appxmanifest">
       <SubType>Designer</SubType>
     </AppxManifest>
+    <Content Include="..\Portable\Images\export.png">
+      <Link>export.png</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="..\Portable\Images\save.png">
+      <Link>save.png</Link>
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="Assets\Fonts\segoeui.ttf" />
+    <Content Include="Assets\Fonts\segoeuib.ttf" />
+    <Content Include="Assets\Fonts\segoeuii.ttf" />
+    <Content Include="Assets\Fonts\segoeuil.ttf" />
+    <Content Include="Assets\Fonts\segoeuisl.ttf" />
+    <Content Include="Assets\Fonts\segoeuiz.ttf" />
+    <Content Include="Assets\Fonts\seguibl.ttf" />
+    <Content Include="Assets\Fonts\seguibli.ttf" />
+    <Content Include="Assets\Fonts\seguili.ttf" />
+    <Content Include="Assets\Fonts\seguisb.ttf" />
+    <Content Include="Assets\Fonts\seguisbi.ttf" />
+    <Content Include="Assets\Fonts\seguisli.ttf" />
+    <Content Include="Assets\Fonts\telerikfontexamples.ttf" />
     <None Include="Windows_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="..\Portable\Images\FullCrop.png">
-      <Link>FullCrop.png</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="..\Portable\Images\AbsCrop.png">
-      <Link>AbsCrop.png</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="..\Portable\Images\PropCrop.png">
-      <Link>PropCrop.png</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
     <Content Include="Properties\Default.rd.xml" />
     <Content Include="Assets\LockScreenLogo.scale-100.png" />
     <Content Include="Assets\LockScreenLogo.scale-125.png" />
